requirements/prod.txt
# Core
pytz==2021.1
Django==3.2.12
gunicorn==20.1.0
django-dotenv==1.4.2
django-3-jet==1.0.8
celery==4.4.7
pyyaml

# For the persistence stores
psycopg2==2.8.6
redis==3.5.3

# Model Tools
django-model-utils==4.1.1
django_unique_upload==0.2.1
django-celery-beat==2.0.0
django-activity-stream==0.10.0

# Rest apis
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
Markdown==3.3.4
drf-yasg==1.20.0
django-filter==2.4.0
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
django-auto-prefetching==0.1.10

# Developer Tools
ipython==7.24.0
django-inlinecss==0.3.0

# Static and Media Storage
django-storages==1.11.1
boto3==1.17.84

# App requirement
feedparser==6.0.10